+++ src/schema/__tests__/generateFroidSchema.fed-v2.test.ts
@@ -1,17 +1,23 @@
 import {generateFroidSchema} from '../generateFroidSchema';
-import {print} from 'graphql';
+import {print, DefinitionNode} from 'graphql';
 import {stripIndent as gql} from 'common-tags';
+import {ObjectTypeNode} from '../types';
 
 function generateSchema(
   subgraphs: Map<string, string>,
   froidSubgraphName: string,
   contractTags: string[] = [],
-  typeExceptions: string[] = []
+  typeExceptions: string[] = [],
+  nodeQualifier?: (
+    node: DefinitionNode,
+    objectTypes: Record<string, ObjectTypeNode>
+  ) => boolean
 ) {
   return print(
     generateFroidSchema(subgraphs, froidSubgraphName, {
       contractTags,
       typeExceptions,
+      nodeQualifier,
     })
   );
 }
@@ -61,7 +67,7 @@ describe('generateFroidSchema for federation v2', () => {
     );
   });
 
-  it('does not propogate miscelaneous directives to the generated id field', () => {
+  it('does not propagate miscellaneous directives to the generated id field', () => {
     const productSchema = gql`
       type Query {
         topProducts(first: Int = 5): [Product]
@@ -402,7 +408,7 @@ describe('generateFroidSchema for federation v2', () => {
           id: ID!
           brandId: Int!
         }
-        
+
         type Product implements Node @key(fields: "upc sku brand { brandId }") {
           id: ID!
           upc: String!
@@ -449,6 +455,73 @@ describe('generateFroidSchema for federation v2', () => {
 
         interface Node {
           id: ID!
+        }
+
+        type User implements Node @key(fields: "userId") {
+          id: ID!
+          userId: String!
+        }
+      `
+    );
+  });
+
+  it('ignores types based on a custom qualifier function', () => {
+    const userSchema = gql`
+      type Query {
+        user(id: String): User
+      }
+
+      type User @key(fields: "userId") {
+        userId: String!
+        name: String!
+      }
+
+      type Todo @key(fields: "oldTodoKey") {
+        oldTodoKey: String!
+      }
+    `;
+
+    const todoSchema = gql`
+      type Todo @key(fields: "todoId") @key(fields: "oldTodoKey") {
+        todoId: Int!
+        oldTodoKey: String!
+        text: String!
+        complete: Boolean!
+      }
+    `;
+    const subgraphs = new Map();
+    subgraphs.set('todo-subgraph', todoSchema);
+    subgraphs.set('user-subgraph', userSchema);
+
+    const nodeQualifier = (node) =>
+      node.name.value !== 'Todo' ||
+      node.directives.filter((directive) => directive.name.value === 'key')
+        .length > 1;
+
+    const actual = generateSchema(
+      subgraphs,
+      'relay-subgraph',
+      [],
+      [],
+      nodeQualifier
+    );
+
+    expect(actual).toEqual(
+      // prettier-ignore
+      gql`
+        extend schema @link(url: "https://specs.apollo.dev/federation/v2.0", import: ["@key", "@tag"])
+
+        type Query {
+          node(id: ID!): Node
+        }
+
+        interface Node {
+          id: ID!
+        }
+
+        type Todo implements Node @key(fields: "todoId") {
+          id: ID!
+          todoId: Int!
         }
 
         type User implements Node @key(fields: "userId") {