+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.2.0] - 2023-06-27
+
+### Added
+
+- Add support for a custom schema node qualifier when generating node-relay
+  schema so users can determine whether or not to include an entity in the
+  generated node-relay schema based on custom criteria
+
 ## [v2.1.0] - 2023-06-01
 
 ### Added